// afn-client/src/app/models/user.ts
export interface User {
  _id?: string;
  username: string;
  email: string;
  password: string;
  date?: Date;
  type?: string;
  firstname?: string;
  lastname?: string;
  cin?: string;
  dateB?: Date;
  profession?: string;
  adresse?: string;
  ville?: string;
  tel?: string;
  gender?: string;
}